Creació de Variables Constants per a modificació del Sistema, i controller,...

See merge request mbotana/projecte4_entrebicis!2

+++ EntreBicis_Backend/src/main/java/cat/copernic/mbotana/entrebicis_backend/entity/SystemParams.java
@@ -28,9 +28,9 @@ public class SystemParams {
     private Double pointsConversion;
 
     @Column(nullable = false)
-    private Double stopMaxTime;
+    private int stopMaxTime;
 
     @Column(nullable = false)
-    private Double collectionMaxTime;
+    private int collectionMaxTime; 
     
 }
